Rename `newton.tests.builders` to `newton.tests.utils`
