+++ newton/_src/solvers/kamino/examples/newton/example_kamino_basic_boxes_nunchaku.py
@@ -17,7 +17,7 @@ import warp as wp
 import newton
 import newton.examples
 from newton.tests import get_kamino_basics_asset
-from newton.tests.builders import basics
+from newton.tests.utils import basics
 
 
 class Example:
